Dev: fixed supports remarks

+++ src/Denominations.java
@@ -1,11 +1,11 @@
 public interface Denominations {
-    long FIVE_HUNDRED = 500;
-    long TWO_HUNDRED = 200;
-    long ONE_HUNDRED = 100;
-    long FIFTY = 50;
-    long TWENTY = 20;
-    long TEN = 10;
-    long FIVE = 5;
-    long TWO = 2;
-    long ONE = 1;
+    int FIVE_HUNDRED = 500;
+    int TWO_HUNDRED = 200;
+    int ONE_HUNDRED = 100;
+    int FIFTY = 50;
+    int TWENTY = 20;
+    int TEN = 10;
+    int FIVE = 5;
+    int TWO = 2;
+    int ONE = 1;
 }
